Changes according to suggestions
