+++ src/MW/exception/MaxLoadExceededException.java
@@ -1,4 +1,4 @@
-package trip.exception;
+package MW.exception;
 
 
 public class MaxLoadExceededException extends Throwable {
