+++ src/MW/service/Car.java
@@ -1,4 +1,4 @@
-package java.service;
+package MW.service;
 
 public class Car {
