+++ src/MW/service/InputExecutiveService.java
@@ -1,4 +1,4 @@
-package java.service;
+package MW.service;
 
 import java.io.BufferedReader;
 import java.io.IOException;
