+++ src/MW/service/InputService.java
@@ -1,4 +1,4 @@
-package java.service;
+package MW.service;
 
 public interface InputService {
