+++ src/MW/trip/Voyager.java
@@ -1,4 +1,4 @@
-package java.pack;
+package MW.trip;
 
 
 public class Voyager {
